+++ tests/daily_model/test_fit_model.py
@@ -17,6 +17,9 @@
    limitations under the License.
 
 """
+
+from pathlib import Path
+
 import numpy as np
 import pandas as pd
 from eemeter.eemeter.models.daily.model import DailyModel
@@ -24,11 +27,15 @@ from eemeter.eemeter.models.daily.data import DailyBaselineData
 from eemeter.eemeter.models.daily.optimize_results import OptimizedResult
 
 
+# Define the current directory
+current_dir = Path(__file__).resolve().parent
+
+
 class TestFitModel:
     @classmethod
     def setup_class(cls):
         # Create a sample meter data DataFrame from the test data
-        df = pd.read_csv("tests/daily_model/test_data.csv")
+        df = pd.read_csv(current_dir / "test_data.csv")
         df.index = pd.to_datetime(df["datetime"])
         df = df[["temperature", "observed"]]
         cls.meter_data = DailyBaselineData(df, is_electricity_data=True)
@@ -75,7 +82,7 @@ class TestFitModel:
 
         # Test that the wRMSE_base attribute is a float
         assert isinstance(fm.wRMSE_base, float)
-        assert np.isclose(fm.wRMSE_base, 18.389335982383994, rtol=1e-3)
+        assert np.isclose(fm.wRMSE_base, 18.39, rtol=1e-2)
 
         # Test that the best combination is as expected
         expected_best_combination = "wd-su_sh_wi__we-su_sh_wi"
@@ -88,11 +95,16 @@ class TestFitModel:
 
         # Test that the error attribute values are as expected
         expected_model_error = {
-            "wRMSE": 16.95324536039207,
-            "RMSE": 16.95324536039207,
-            "MAE": 13.38096518529209,
-            "CVRMSE": 0.32064123575928577,
-            "PNRMSE": 0.270778281497731,
+            "wRMSE": 16.95,
+            "RMSE": 16.95,
+            "MAE": 13.38,
+            "CVRMSE": 0.3206,
+            "PNRMSE": 0.2708,
+            "wRMSE": 16.95,
+            "RMSE": 16.95,
+            "MAE": 13.38,
+            "CVRMSE": 0.3206,
+            "PNRMSE": 0.2708,
         }
         for k in expected_model_error:
-            assert np.isclose(fm.error[k], expected_model_error[k], rtol=1e-3)
+            assert np.isclose(fm.error[k], expected_model_error[k], rtol=1e-2)